+++ packages/link/src/link.android.js
@@ -2,9 +2,10 @@ import React from "react";
 import { TouchableNativeFeedback, View, ViewPropTypes } from "react-native";
 import PropTypes from "prop-types";
 
-const Link = ({ children, linkStyle, onPress }) => (
+const Link = ({ children, linkStyle, onPress, disabled }) => (
   <TouchableNativeFeedback
     delayPressIn={0}
+    disabled={disabled}
     onPress={onPress}
     useForeground={TouchableNativeFeedback.canUseNativeForeground()}
   >
@@ -18,11 +19,13 @@ const { style: ViewPropTypesStyle } = ViewPropTypes;
 
 Link.propTypes = {
   children: PropTypes.node.isRequired,
+  disabled: PropTypes.bool,
   linkStyle: ViewPropTypesStyle,
   onPress: PropTypes.func.isRequired
 };
 
 Link.defaultProps = {
+  disabled: false,
   linkStyle: {}
 };
