+++ packages/link/src/link.js
@@ -2,8 +2,8 @@ import React from "react";
 import { TouchableOpacity, ViewPropTypes } from "react-native";
 import PropTypes from "prop-types";
 
-const Link = ({ children, linkStyle, onPress }) => (
-  <TouchableOpacity onPress={onPress} style={linkStyle}>
+const Link = ({ children, disabled, linkStyle, onPress }) => (
+  <TouchableOpacity disabled={disabled} onPress={onPress} style={linkStyle}>
     {children}
   </TouchableOpacity>
 );
@@ -12,11 +12,13 @@ const { style: ViewPropTypesStyle } = ViewPropTypes;
 
 Link.propTypes = {
   children: PropTypes.node.isRequired,
+  disabled: PropTypes.bool,
   linkStyle: ViewPropTypesStyle,
   onPress: PropTypes.func.isRequired
 };
 
 Link.defaultProps = {
+  disabled: false,
   linkStyle: {}
 };
