+++ packages/star-button/README.md
@@ -1,6 +1,22 @@
-# StarButton
+# Star Button
 
-<!-- Add the description of the package here. -->
+A star button component which is clickable and exposes three states: `initial`, `selected`, `disabled`
+
+- `StarButton` - wraps any star svg and is clicable. Can be disabled.
+- `starStates` - enum for the button states.
+
+## How to use
+
+```js
+import StarButton, { starStates } from "@times-components/star-button";
+
+// works this way
+<StarButton onPress={this.onStarPress} starState={starState} />
+
+```
+
+`starState` prop controls the state of the star. It changes colours and opacity based on the different state.
+`Diabled` would not allow the star to be clicked.
 
 ## Contributing
 
@@ -22,24 +38,16 @@ yarn lint
 
 ## Testing
 
-Testing can be done on each platform individually
+This package uses [yarn](https://yarnpkg.com) (latest) to run unit tests on each
+platform with [jest](https://facebook.github.io/jest/).
 
 ```
+yarn test:all
 yarn test:android
 yarn test:ios
 yarn test:web
 ```
 
-Or the tests for all platforms can be run
-
-```
-yarn test:all
-```
-
 Visit the official
-
-<!-- Add the storybook link here. -->
-
-## Future
-
-<!-- Add details of future development here. -->
+[storybook](http://components.thetimes.co.uk)
+to see our available link templates.
