+++ packages/star-button/package.json
@@ -37,14 +37,14 @@
   "homepage": "https://github.com/newsuk/times-components#readme",
   "devDependencies": {
     "@thetimes/jest-lint": "*",
-    "@times-components/eslint-config-thetimes": "*",
-    "@times-components/jest-configurator": "*",
-    "@times-components/storybook": "*",
-    "@times-components/webpack-configurator": "*",
+    "@times-components/eslint-config-thetimes": "0.8.9",
+    "@times-components/jest-configurator": "2.3.4",
+    "@times-components/storybook": "3.4.5",
+    "@times-components/webpack-configurator": "2.0.15",
     "babel-cli": "6.26.0",
-    "eslint": "4.19.1",
+    "eslint": "5.9.0",
     "jest": "23.3.0",
-    "prettier": "1.8.2",
+    "prettier": "1.14.3",
     "react": "16.5.2",
     "react-dom": "16.5.2",
     "react-native": "0.55.4",
@@ -53,8 +53,9 @@
     "webpack-cli": "2.1.4"
   },
   "dependencies": {
-    "@times-components/styleguide": "3.26.3",
     "@times-components/icons": "2.8.0",
+    "@times-components/link": "3.2.30",
+    "@times-components/styleguide": "3.26.3",
     "prop-types": "15.6.2"
   },
   "peerDependencies": {
