+++ packages/star-button/src/star-button.js
@@ -1,8 +1,8 @@
 import React, { Component } from "react";
-import { TouchableOpacity } from "react-native";
 import PropTypes from "prop-types";
-import { colours } from "@times-components/styleguide";
+import Link from "@times-components/link";
 import { IconStar } from "@times-components/icons";
+import stars from "./styles";
 
 const starStates = {
   disabled: "disabled",
@@ -13,25 +13,6 @@ const starStates = {
 class StarButton extends Component {
   constructor(props) {
     super(props);
-
-    const stars = {
-      disabled: {
-        fillColour: "none",
-        opacity: "0.4",
-        strokeColour: colours.functional.secondary
-      },
-      initial: {
-        fillColour: "none",
-        opacity: "1",
-        strokeColour: colours.functional.secondary
-      },
-      selected: {
-        fillColour: colours.functional.action,
-        opacity: "1",
-        strokeColour: "none"
-      }
-    };
-
     this.state = { star: stars[props.starState] };
   }
 
@@ -41,14 +22,14 @@ class StarButton extends Component {
     const disabled = starState === starStates.disabled;
 
     return (
-      <TouchableOpacity disabled={disabled} onPress={onPress}>
+      <Link disabled={disabled} onPress={onPress}>
         <IconStar
           fillColour={fillColour}
           height={height}
           opacity={opacity}
           strokeColour={strokeColour}
         />
-      </TouchableOpacity>
+      </Link>
     );
   }
 }
