+++ packages/star-button/src/styles/index.js
@@ -1,3 +1,21 @@
-const styles = {};
+import { colours } from "@times-components/styleguide";
 
-export default styles;
+const stars = {
+  disabled: {
+    fillColour: "none",
+    opacity: "0.4",
+    strokeColour: colours.functional.secondary
+  },
+  initial: {
+    fillColour: "none",
+    opacity: "1",
+    strokeColour: colours.functional.secondary
+  },
+  selected: {
+    fillColour: colours.functional.action,
+    opacity: "1",
+    strokeColour: "none"
+  }
+};
+
+export default stars;
